Clean up code with the aid of comments given by peers
Some coding standard violations such as not leaving a line in JavaDocs
between the description and the @params and @return statements

Made changes accordingly to fit the guidelines for standardization,
to ensure that JavaDocs written adhere to these standards and are of
high quality, and to maximize learning. Names of some variables were
also changed to better represent what items they were actually storing
to avoid any misconceptions.