Invalid events are accepted by Duke, where the end date is earlier than start date.
An event with an end date canonot be earlier than the start date, this is to ensure that Duke catches this error and doesn't accept this invalid input. We want to ensure the correctness of our Tasks in Duke.

Event constructor will throw a DukeException if such an input arrives, by utilising the LocalDateTime isBefore() method.

Event holds the LocalDateTime information, hence one way I thought was to throw the Exception within the Event Class.