`.chattime.storage.Storage ` `loadData()`: need checking on loaded data                                            length `.chattime.Chattime` `getResponse(String input)`: need checking on                                                   processed commands
Lengths of data are specified for different types of tasks.
Processed commands should not be null.

Let's add assertions on lengths of loaded data according to types.
Let's add assertions that cmd should not be null.

The lengths of loaded data and the parsed commands are not accessible
by users, they need to have assertions to ensure that the programs are
running in a normal and good condition.

Assertion error messages are added for clarification.