Allow generation of working JAR file
Previously shadowjar for Gradle would exprot a JAR file that keeps
crashing. This was caused by the resource not being found because
of the file name. Resources under out extends the file with .PNG
instead of .png which caused the error.

New JAR file will properly load resources by looking for images as
.PNG extension.

Future management of resources must be carefully handled to consider
file extensions too.