Refactor Duke class to adhere to code quality standards
Duke methods does not adhere to code quality standards.

Poor code quality can result in readers not being able to understand the code whilst making it more prone to bugs and errors.

Usage of blank lines in order to implement logical structuring of code in the getResponse method and the removal of redundant method calls in the same method.

By laying out the code such that it adheres to logical structure, this makes it very clear for readers to easily tell at a glance what we are doing. Furthermore, the removal of redundant code can reduce the complexity of code being read by readers allowing them to better understand our intentions as the developer.