A-CodeQuality
There are existing code that are not of the best code quality. An example
include the handling of invalid inputs. Error handling is done in individual
command classes causes issues when customising error messages.

Using the SLAP principle would allow the code to be more readable and
faciliate future changes.

Extracting out the error handling from individual command classes into
the parser class allows the error handling to be done in a centralised
location.