Add udpate feature
Created UpdateCommand class which executes update method in TaskList
class. Command format is "update <index> <new description or /by, /from
or /to> <new date and time if previouis field was /by, /from or /to>".

Tasks can now be updated instead of having to delete it and then adding
it with the changes again.