Change access modifiers and variable names
The variables in a class with no intended inheritance should be set to private to prevent unintended exposure of the subclass. There are also instances of inappropriate variable names to represent a plural object.

Improving the access modifier from protected to public, and representing the variables in their correct grammar form improves code quality.

This improves readability so that the code can be easily maintained and referred.

Let's,
* change the access modifiers of variables in classes with no child classes to private
* update variable names to better suit its object type - singular or plural, especially if there are any changes made to its implementations