Add reminder command
There is no reminder command for users to see upcoming deadlines and
events.

A reminder command is useful for users to check what to do soon, so that
they do not have to check the whole list of tasks.

Let's add a reminder command that will display the deadlines that end
and events that start within 2 days from the time that the command is
entered. A todo will not be reminded about as it does not have a date
that it needs to be finished by.