Change UI
The text is slanted, and can be blocked by text input box when the
text is too long.

Straightening it can allow long text to be viewed.

Let's straighten the label that has the text.

It is preferable to allow all text to be seen if the user adds many
tasks with long descriptions and displays them.