Introduce the use of java assertions to verify invariants
Several invariants, such as the existence of key objects and the
correctness of arguments when calling certain functions are to be
verified before these functions can be safely called. Otherwise,
there is a chance of undefined or unexpected behavior. This commit
uses java assertions to verify some of these invariants.