Add java assertions to code, add IntParseDukeException class
Some parts of the code relies on the programmer entering
the correct input argument in order for the code to not
exhibit any unwanted behaviour.

Let's use java assertion to check for such unwanted
behaviours caused by the programmer's wrong input.

Exceptions may also be thrown while trying to parse
a string input that is not an integer for the mark,
unmark and delete command in Duke.

Let's handle this exception by creating an
IntParseDukeException that gets throwns when such
an error occurs.