A-CodeQualiy: Extract input validation check
The original version caused clutter of if condition in the execute
function resulting in
* poor readability
* unneccessary repeats

New class `InputValidation` is created to handle all validation and
cleaning of the request from the user before parsing it to add command.

It is done this way as three commands, respectively `mark`, `unmark`
and `remove` have the same input validation check. Hence, to follow the DRY
principle, it is extracted out so when needed can simply call the
designated function `normaliseIndexRequest`. Also, it improves the
cleanliness of code.