A-Streams: add streams to `find` method
Previous version uses for loop as list may get longer and for loop might
have a longer run time.

Task list is being transformed to streams and use `stream::filter` to
filter using the keyword provided. The return result is all task with
keyword collected in an array list.

This is done this way to accomodate for further extension of the list.