Use assert to document important assumptions
When each input is being parsed and returned as a hashmap,
the keys should contain keywords required for the relevant command.

By recording these assumptions, it serves as a check to ensure that
the keys for the relevant command appear in the hashmap for execution.
This way, the command should execute and output as expected.