Add assertions to test code
No checks are done to ensure that a response is formed when user enters input or that
a file path exists

Asserting that file path exists ensures that creating a File object with the file path will not
throw an error. Asserting that the response is not an empty String ensures that there is no
error in parsing user commands and having Duke respond to the command.

Using assertions is desirable because it can indicate any errors made by the coder.