Add reminder to Duke and LoadDukeException class.
Users have to manually look at their to do list in
Duke in order to find any tasks that is about to happen.

Let's create a reminder that appears everytime Duke starts
up to remind users that the given task is about to happen.

The number of days to the task to be compared to is set at
10 days. The ability to change this day will be implemented
in the future.

A new DukeException class, LoadDukeException has also been
created to deal with errors arising from loading the file
from the given path to Duke.

Code has been refactored to make the reminder function work
and to throw and catch LoadDukeException.