Add specific help command for different commands
Help message only displayed when app starts, and all of the information
is in that 1 message.

There is a need to scroll all the way to the beginning if the user
forgets how to use a command. It is also very cluttered with
information on how to use the commands.

Let's add specific help commands like help [command], and change
help prompt on app start to something more concise.

Keeping the starting help prompt tidy will allow users to quickly
navigate to what they want, and help requests for specific commands
will allow more detailed explanations and examples on how to use
the command.