Add ReminderCommand.java to create reminder function Edit the following classes to create the reminder function: - Deadline.java - Duke.java - Event.java - Parser.java - Task.java - ToDo.java - FindCommand.java
Notes:
Point 1:
The newly implemented reminder function via ReminderCommand tracks
all Deadline and Event tasks along with their completion status.
Tasks are being checked against the system date and time.
Therefore, ENSURE the system is SET to the CORRECT date and time.

Point 2:
ToDo tasks are also tracked but via their completion status and not
date time since they are not allocated with any!
It is ASSUMED that all ToDo tasks are to be done ASAP!