Add an assertion in Parser class
An assertion added in the handleInput method ensures the task is marked
correctly when given a command to mark or unmark a particular task as
done.

Let's insert an assertion to check if the task given is marked or
unmarked as told to do.

Using assertions can detect potential bugs in code when an assertion
failure is detected