Add remind feature, abstract LocalDateTime methods, improve code quality
Duke cannot remind a user if a task is due on the current day. There
are also magic symbols to split storage text and a lot of methods
using LocalDateTime is in the Task class.

We want to add a new feature called remind and we should encapsulate
LocalDateTime in a more relevant class.

Let's,
*Add remind command that reminds user on events and deadlines due
today
*Refactor LocalDateTime methods to a new Time class
*Make minor changes to improve code quality like removing magic
symbols