BCD-Extension: Add C-Update as a command to classes
BCD-Extension requires choosing at least one extension to add a function into the program

Implementation of updateTask method with parsing of index and input string to update the item in the list corresponding to its index. However, user are required to input the whole instruction in order for the program to return the correct outputs.

Implementing an update method helps to improve the user experience as users would find it easier to just update the item with one command, instead of deleting and re-adding of the item onto the list. This is also to retain the original list sequence.