Add new Edit Command
Allows the user to indicate which Task to edit. The number of fields
required is at least 1, with the consideration that the user might
only want to change a subset of attributes.

To be more specific, after an edit, a new Task is created and replaces
the pre-existing (to be edited) task in the TaskList. This was
implemented as so to maintain some immutability of Tasks.