New look for Duke and User
The images for Duke and User are the default images in the tutorial.

The default images are not very appealing and not suitable for this app.
The images also remain the same regardless of the state of the app.

Lets change the images to be more appealing.
Lets also dynamically alter the user image shown depending on
how many tasks are in the current task list.

This will provide a more enjoyable experience for the end user.