Add assertions to ensure code correctness
Assumptions about the state of the program at any point
during program execution should be ensured to hold true.

Let's use assertions to help us achieve this. We ensure
that the following behaviours:
* Resulting response string from a valid command should
not be empty.
* Memory Path Array supplied to the Storage constructor
should not be empty.
* Number of results arising from a search query cannot
be negative.
* Response string obtained by Duke from the Parser cannot
be null

Using assertions liberally is reccommended to reduce the occurences of bugs due to human error.