Improve code quality and fix a bug
There are still some room of improvement for the code quality. There is
a bug where the tag will duplicate.

The bug can cause major issues.

Let's add some more inline methods and remove the bug by resetting the
tag string for each element in the writing method for the storage.

Best way to fix the current bug.