Modify find method to utilize Stream
Current implementation loops through the list of tasks to check if
its description contains the argument, then add it into list of
found tasks, then prints out based on the found tasks.

This can be efficiently implemented with Streams.

Printing all tasks in lists can also be modified to use Streams
in the future.