Add assertions
Kuromi is able to add, delete, get data from storage.

However, there are invariants that should be followed by the
application.

To check if the application follows the invariants, we can add
assertions to ensure that the app follows the rule (e.g. adding a task
increases the task list size by 1, deleting a task decreases the task
list size by 1, and the task list size of a user obtaining tasks from
the storage is the same as the number of tasks in the storage file.