Add query for free time slots in user schedule
Users were previously unable to query for free times in their list of
tasks.

To resolve this issue, adding a new feature allowing users to state a
duration and the bot automatically finds a free slot if it exists.

Introduce FreeTimeBlock class to represent free timeslots with start and
end timings.

Implement getFreeTimes() method in the TaskManager class to find free
slots based on the user's schedule.

A different String representation of FreeTimeBlock is generated depending on
whether it is bounded or unbounded.

This update allows for more efficient time management, and improves user
experience.