Command: improve assert messages
With the recent change to text parsing, it is
more likely for assertions to be violated. This is
especially true for the runCommand functions, where
we want to ensure only the correct Commands are
accessing them.

Let's clarify the assert messages of the runCommand
functions.

If we remove ambiguity from which is which, it will
be easier to debug.