Added new extension: Undo
Users may want to undo the very last command.

Undo command has been added to Duke.
If the last command is only for listing/finding purpose, it will not be able to undone as these commands do not change any data that Duke has.

Users can now undo the last command, given that it is applicable.