Align with the Code Quality guidelines
The Parser class contains only one method (parse) that contains the
processes to split the input into description, deadline, etc. for
each command.

This makes the method to be too long (more than 30 lines).

Therefore, I made a method for each command to separate the processes.
The event command might be still too long (around 40 lines), I didn't
find a way to shorten the number of lines for the method.

Furthermore, I changed the execute method for every child classes of
Command class, so that it aligns with SLAP. I also changed the code
in parser method to align with SLAP