The current implemention does not use any assertions.
Assertions should be implemented because they can be a useful tool for improving the quality, clarity, and maintainability of the code.

Some assertions are being added to Storage.java and Duke.java.

It is done this way because without them, there is nothing checking if the inputs are missing or not which could potentially lead to run time errors.