Refactor code and move class methods
Error messages were not in each classes' static variables. To
comply with coding standards and have greater readability,
error messages and reply statements have been moved to static
final attributes in each class.

Previously, each executed Command returns a simple string. With
consideration for the need for greater detail in a response,
a flag isSuccess has been added. The flag serves as an indicator
if an executed Command should have changes be written in. This
eliminates the risk of a failed Command execution resulting in results
being saved.