Code Quality: Shorten EventResponse::exec
EventResponse::exec is well over 30 lines

The chunk of code that is responsible for parsing and checking the inputs for errors are extracted out into another method

Long and verbose methods are very unreadable and difficult to debug.
For example, if the method is really long, a bug can come from either the parsing of inputs or the creation of a new Event. However, now that it is separated, at a glance, one can tell where the bug is coming from.