Add assertions to Duke
In many methods, it is assumed that certain variables or objects
are of a certain state.

By using assertions, code quality can be improved to make those
assumptions a guarantee.

Assumptions make it difficult to debug if the error
stack trace is very long.

It is important to hold assumptions in current state as otherwise
there will be a need to check for many things at all methods.

Let's improve code quality and make debugging easier by adding
assumptions.