Add Assertions
There are portions of code which expect arguments to be non-null, or
call methods which are expected to have certain behaviours.

If these code elements encounter incorrect arguments or method
behaviour, they may throw difficult-to-diagnose Exceptions or fail
silently.

Let's add assertions to these portions of code to make it easier to
diagnose such situations, if they arise due to bugs elsewhere in this
project or in external Java modules.

Using assertions allows a visible and clear error to be thrown when an
unexpected situation arises, with the affected source line indicated and
a relevant message of choice included.