Bugfix: Remove incorrect assert
Assert to ensure that something is written to file.

We need to consider the situation when user deletes all tasks.
Then task list will be empty (is valid state), meaning there is
nothing to write to file. This falsely fails the assert.

Let's remove the assert.

In general, we should writing a blank to a file.