Adds assertions to Duke
Currently, no assertions are used in Duke's code.
Assertions are useful a useful lightweight tool that ensures that
assumptions are kept and not violated.

Let's add some assertions to Duke's code. Particularly in places that
assume values are not null or not empty.

This should make Duke less error prone and ensures assumptions are
always kept no matter the version of Duke.