Refactors DatedTasks to improve code quality
The compareTo function in DatedTasks has multiple nested if else that
determine the logic of how DatedTasks are sorted.

This is bad as this makes the logic difficult to follow and understand.

Let's bring out the nested ifs and remove the else by relying on the
return statement to stop the flow of instruction.

This makes the logic of the compareTo function easier to follow and
imporves the code quality.