Adds JavaDocs to DialogBox
There is no JavaDocs for the public methods of getUserDialog and
getDukeDialog.

Public methods should always have JavaDocs unless they are getter/
setters or have inherited JavaDocs.

Let's add the JavaDocs to these two methods.

This will ensure that the behaviour of the method is understood by the
user of the method.