Refactors toCompare method in Task.java
The toCompare method has multiple nested if else that determine the
logic when Tasks are sorted.

These multiple nested if else makes the logic difficult to understand
and the consequent behavour difficult to predict.

Let's bring out the ifs and remove the else.

This makes ensures that the logic is better understood and the behavour
is more easily predicted.

The elses can be removed as we can use the return statements to stop the
instruction flow.