Adds CI to Duke
Currently, Duke is not continuously integrated for any updates pushed or
merged in pull requests.

This means that it is possible for a version of Duke to not work and
have devlopers/users not realise.

Let's set up a CI for Duke.

This will automatically make sure that Duke is compilable in gradle and
therefore inform developers/users if the latest changes made have broken
Duke.