Command Classes: Add assertions
execute() function within each instances of commands perform
certain actions and contains toResultString() subroutine to
convert actions done to printable string for the UI to show
the user what was done.

There is strong coupling between the two functions
as the context of the command is being stored as attributes within
the command classes by execute()) (e.g. taskList, deletedTask)) and is
used within the toResultString() subroutine.

Creating assertions within toResultString() subroutine
will ensure that the result of the execution of command is returned as a
string and is correct by ensuring the assumption that the context is
stored by the command classes before execution of relevant action, after
which the result can be returned in a string format.