MainWindow AddCommand, DeleteCommand, FindCommand, UnmarkCommand do no have assertion to check if the output messages are correct
The lack of assertion can leave inaccurate output messages unchecked

Adding assertion can ensure that the output messages are accurate

Let's java assertion to each class with unique error message should the assertion fail. For example, if the output message of FindCommand is inaccurate, a message "wrong find message" will be output.

Additionally, a get() method is added to TaskList class to facilitate addition of assertion to the above classes.