Class Storage
there are redundant code that is no longer useful. For example, the Status attribute and some output messages

the name of the method "update" is not meaningful enough.

readers of the code might get confused when reading the code

removing the unused attribute and output messages increases the readability of the code

Let's remove the redundant lines and refactor the method "update" to "updateLocal"

Class Duke

method "getLoadStatus" is no longer used because the Status attribute in Storage is removed

removing this method prevents error from occurring

Let's remove the unused method getLoadStatus()