Classes Duke, Command, DeleteCommand, UnmarkCommand, AddCommand, ByeCommand, FindCommand, MarkCommand, ListCommand, CommandType, DukeException, Parser, Storage, Deadline, Event, Task, Todo, TaskList, Ui.
Header comments are not in accordance with the format specified in the coding standard.

Comments in the wrong format makes it harder for the readers to understand various methods

Let's change the problematic comments in accordance with the format specified in the coding standard

Additionally, unused method "readCommand" in Ui class is removed.