Add archive feature
Duke does not have an archive feature

To enhance the user experience, as time goes by, the tasklist might be filled with a lot of tasks and it will be messy and hard for the user to keep track of his tasks.

Duke now has an archive feature, which can be accessed by typing "Archive" into the input box. Archive will delete all of the tasks in the tasklist as well as the data in data/duke.txt. The deleted data will be archived in a new file named "current datetime eg. 021120232053" and saved in the directory "data".