Storage.java: added constants for magic strings and numbers
Magic strings and numbers are present in Storage.java.

Having these magic strings and numbers may cause confusion as they may
be difficult to interpret, especially to other people who may read or
work on the code in the future.

Defining these strings and numbers as constants makes them more
descriptive. This improves readability of the code.