Fix minor bugs
Storage class prints to the output console directly when it encounters
an IOException when writing to the test file.

This is not good behaviour as anything to be printed to the output
consolse must be done by the TextUi class.

So, we will make the Storage object throw a DukeException if it
encounters the IOException, for it to be handled by functions
higher in the call stack.

This will ensure that all possible checked exceptions are
handled at a single place.