Change Ui to fit the coding standards
There was text present that contained the default String drawing of Duke, which was relevant only in CLI, but has since become redundant with the new GUI.

This currently unused text is thus deleted in accordance with the coding standard to remove unused and outdated code.

Let's try to find other instances in the code which could be simplified and refactored.