feat: Add assertions for argument of "mark", "unmark" and "delete"
These three commands (i.e. "mark", "unmark" and "delete") receives an
index as their argument, which refers to the index of the task in the
TaskList.

As TaskList is implemented with an ArrayList, to avoid any errors
resulting from index-referencing, it is important to ensure that
the index argument is:
* an integer
* within the range of the ArrayList

Let's
* create a private method in Parser to assert the index to be an
integer
* create a private method in TaskList to assert the index to be within
ArrayList's range