Fixes GUI issues.
GUI does not handle long outputs properly and wraps text. This leaves user with missing information.
Responses are now `Text` instead of `Label` to work around this issue.

`Text` as response is more suitable as `Label` are meant for controls. `Text` also allows dialog boxes to resize according to the length of the response.