Uses `assert` to test at various points of the code.
Assert is used in Parser class to check the assumption that methods parses inputs correctly. This is done by checking that the number of arguments returned after parsing equals the expected number of arguments. Assert is also used in Command classes. As retrieving arguments is hardcoded with the expected index of the arguments, it should be asserted that there are only the expected arguments present.