Separate the handling for each command into individual classes
The user's input needs to be parsed and the program will subsequently
execute the methods corresponding to the command identified.

Combining the parsing and the execution of different commands in one
method makes it long and cluttered.

Hence, the code segment for each command is extracted and placed into
separate methods in their respective classes.