Improve the GUI
Improve the GUI to make it more polished.

Tweak the GUI to match the asymmetric nature of the conversation: As
the conversation is between the user and the app (not between two
humans), it is asymmetric in nature. Accordingly, it makes sense not to
display both sides of the conversion in the same visual format.

Tweak padding, fonts, colors, alignments to make the GUI more pleasing
to look at.

Allow resizing of the Window, and ensure the content resize
appropriately as the Window changes size.

Rounded profile pictures