Add Genie greeting message
The application boots up a blank window that waits for user's input.

Receiving Genie's greeting first gives a more 'complete' appearance and
ensures that the application has started successfully.

Follows weak SLAP principle as there is no abstraction of the greeting
message due to FXML load exception when done so.