GitHub Actions: Set up CI for project
On every push/merge, there is a chance of regression when the code
changes.

This could be catastrophic if left undetected for multiple commits which
might even build on the broken functionality.

Let's set up a simple continuous integration script that runs on GitHub
Actions to ensure that such regressions are caught and rectified as
early as possible, if any.

This is because GitHub actions allow us to do automated tesing on every
push/merge, which helps as manual testing is tedious and may not be as
consistent as automated tests.