Add Assertions to the Ui, MainWindow and Storage
The different classes have underlying conditions for the code to run.

Having assertions make notifies the developers when such conditions are false.

Therefore, I add assertions to the Ui, MainWindow and Storage class, to check for whether the input command is null, whether the GUI have pictures and when the storage does not load from a txt file in the correct template.