Improve code quality and readability in Parser
There was a complicated "if else" expression to determind the types of task as there were a few or expressions to check for all types of task.

Breaking down the "if else" expression into their own individual expressions improved the readability and code quality.

This ensures that there are seperate conditional expressions to consider each variation of task added, allowing for easier understanding of code.