Add functionality for the user to add a tag to a task when creating it
The tasks in the task list cannot be categorised. This functionality allows users to categorise their
tasks by the tag name.

Adding an attribute of type Tag to the abstract class Task allows the tag to be displayed when the
tasks are listed and the users are now able to get a list of all the tasks under the same tag.