Find command: ability to find all tasks of a particular type
User should have the ability to find all the todos/deadlines/events in
the task list.

Let's add functionality so the user can just type "find todo" to
list out all the todos in the list.

Makes it easier for users to filter out tasks of one type if the
list is too long.