Improve CodeQuality
There are several methods that use complicated expressions.

`mark`, `unmark`, `deadline` and `delete` methods in Parser class has 2
checks in an if condition.
Abstracted the checks to 2 variables to enhance readability.

`event` method in the Parser class has 3 checks in the if condition.
Abstracted the checks to 3 variables to enhance readability.

As a step forward, let's abstract complicated checks in the conditions.