Add a few catch statements in Parser and edited the error message thrown.
Added catch statements to catch exceptions because they could be thrown when user gave missing commands.

Exceptions that occur when there is a missing time/date or wrong format given will be caught and I will throw a Missing DukeException.

It is done this way since this error is specific to a missing command.

The error message for Missing exception is also improved as I realised the missing command is also used when there is no information on the task name given