parse method in Parser class: Abstract methods to follow SLAP
Parse method is too long, it does not follow SLAP.

Long methods causes confusion and not friendly looking.

Abstracting the specifics helps to improve the level of
complexity within the method.

Let's abstract the specific details to enhance SLAP of the parse method.
It will become more friendlier and less complex.