Implement assertions to Storage and TaskList files
Developers will not know the correctness of the assumption that the task
is not null, or the size of the list of tasks is negative.

Neglecting these assumptions can cause severe errors and even crash of
chatbot.

Let's use assertions to ensure that these assumptions are checked.

This prevents severe errors from occuring to the users, and even
crashing of chatbot.

Assertions are used to ensure our test conditions are true.