Abstract methods for cleaner code
Long methods such as run() in Duke class and manual checking of user's
input for certain words resulted in arrow-head code and overly long
methods.

Manual checking of input was abstracted into more methods in the Parser
class, with a checking of valid integers. Code is cleaner and easier to
follow.