Implements a simple `sort` command
Simple sort feature is implemented. Currently, it is designed to accept a `key` as an argument to sort the list of tasks. This mutates the original task list and therefore, affects the indexes of the next operations.

SortCommand constructor takes in `key` of type String. `Key` is used to return a `Comparator<Task> object` which is consumed by the TaskList to compare tasks in the task list.