Add Summary function
The summary function invokes whenever the app is booted up and
storage has been loaded. It provides an overview of the tasks that
the user has, and is sorted in a neat fashion for easy reading.

A view of the summary can be invoked at anytime by the user through
the 'summary' command.

However, it is to be noted that the summary command does not give the
user any information that allows them to access a certain Task (to
modify, mark, or delete). This might cause some confusion, but
is mitigated by the closing message at the end of a summary.