Added Notes functionality
The previous version of Duke only allowed Users to store Data in Duke if
the data was in the form of a To Do item, an event or a deadline.

Adding the Notes functionality will enable users to better store data in
Duke.

As a step toward better coding quality, all new methods have JavaDoc
whereever necessary and the new functionality is put in a new package.