Friendlier Syntax: Make input required more user friendly
Commands have to be spelt in full.

A misspelling of the first word of the whole input will cause an error
to be received.

A shorter input will make it more flexible for the user to communication
with the chatbot on the instruction that they want the chatbot to
execute.

Users now can use a single letter to execute the type of command to the
chatbot. This reduces room for misspelling.