Adds ability to sort items under `C-Sort`
Duke is now able to sort tasks in the task list chronologically (from nearest upcoming task to the furthest upcoming task) and sort lexicographically by task description.

This is done by implementing a subpackage `taskcomparator` under tasks. `NameComparator` compares the description of the Task. `DateComparator` compares the dates of the tasks if Task has no date associated to it, Task is sorted lexicographically.

As a design choice, only the `start date` is used to compare between tasks as we are assuming that the user wishes to know which task will come up the next immediate.