Handle all exceptions and error handling, improve code quality, add personality.
All known exceptions will not be printed onto console anymore.

Exceptions were printed onto the console and the user would not know
what actions led to an error.

Handle all exceptions and print reason for particular error onto
output pane in GUI.

This will make it known to the user that they have performed an action
that is not appropriate.