Handle exceptions due to date/time parsing
Parsing of date/time is not handled and expects user to always
provide date/time in the correct format.

Fatal if not followed. Wrong format will case the program to crash.

Let's finally handle the exception by prompting user to re-type
using the correct format, and not crash the program (fatal).

Also conform to coding styling.